refactor[User] : refactor packages
and other minor fixes
