--- src/main/java/com/keepyuppy/KeepyUppy/global/config/OAuth2Config.java
@@ -1,51 +1,0 @@
-package com.keepyuppy.KeepyUppy.global.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.oauth2.client.registration.ClientRegistration;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.core.AuthorizationGrantType;
-import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
-
-@Configuration
-public class OAuth2Config {
-
-    @Value("${kakao.client-id}")
-    String clientId;
-
-    @Value("${kakao.client-secret}")
-    String secret;
-
-    @Bean
-    public ClientRegistrationRepository clientRegistrationRepository() {
-        return new ClientRegistrationRepository() {
-            @Override
-            public ClientRegistration findByRegistrationId(String registrationId) {
-                if ("kakao".equals(registrationId)) {
-                    return kakaoClientRegistration();
-                }
-                return null;
-            }
-        };
-    }
-
-    private ClientRegistration kakaoClientRegistration() {
-
-        // TODO: use yml files only instead of hardcoding
-        return ClientRegistration.withRegistrationId("kakao")
-                .clientId(clientId)
-                .clientSecret(secret)
-                .clientAuthenticationMethod(ClientAuthenticationMethod.CLIENT_SECRET_POST)
-                .authorizationGrantType(AuthorizationGrantType.AUTHORIZATION_CODE)
-                .redirectUri("http://localhost:8080/login/oauth2/code/kakao")
-                .authorizationUri("https://kauth.kakao.com/oauth/authorize")
-                .tokenUri("https://kauth.kakao.com/oauth/token")
-                .userInfoUri("https://kapi.kakao.com/v2/user/me")
-                .userNameAttributeName("id")
-                .scope("profile_nickname", "profile_image")
-                .clientName("Kakao")
-                .build();
-    }
-}
-
